More 64-bit

+++ src/sbbs3/exec.cpp
@@ -195,76 +195,76 @@ int32_t * sbbs_t::getintvar(csi_t *bin, int32_t name)
 			sysvar_l[sysvar_li]=useron.leech;
 			break;
 		case 0x7c602a37:
-			return((long *)&useron.misc);
+			return((int32_t *)&useron.misc);
 		case 0x61be0d36:
-			return((long *)&useron.qwk);
+			return((int32_t *)&useron.qwk);
 		case 0x665ac227:
-			return((long *)&useron.chat);
+			return((int32_t *)&useron.chat);
 		case 0x951341ab:
-			return((long *)&useron.flags1);
+			return((int32_t *)&useron.flags1);
 		case 0x0c1a1011:
-			return((long *)&useron.flags2);
+			return((int32_t *)&useron.flags2);
 		case 0x7b1d2087:
-			return((long *)&useron.flags3);
+			return((int32_t *)&useron.flags3);
 		case 0xe579b524:
-			return((long *)&useron.flags4);
+			return((int32_t *)&useron.flags4);
 		case 0x12e7d6d2:
-			return((long *)&useron.exempt);
+			return((int32_t *)&useron.exempt);
 		case 0xfed3115d:
-			return((long *)&useron.rest);
+			return((int32_t *)&useron.rest);
 		case 0xb65dd6d4:
-			return((long *)&useron.ulb);
+			return((int32_t *)&useron.ulb);
 		case 0xabb91f93:
-			return((long *)&useron.dlb);
+			return((int32_t *)&useron.dlb);
 		case 0x92fb364f:
-			return((long *)&useron.cdt);
+			return((int32_t *)&useron.cdt);
 		case 0xd0a99c72:
-			return((long *)&useron.min);
+			return((int32_t *)&useron.min);
 		case 0xd7ae3022:
-			return((long *)&useron.freecdt);
+			return((int32_t *)&useron.freecdt);
 		case 0x1ef214ef:
-			return((long *)&useron.firston);
+			return((int32_t *)&useron.firston);
 		case 0x0ea515b1:
-			return((long *)&useron.laston);
+			return((int32_t *)&useron.laston);
 		case 0x2aaf9bd3:
-			return((long *)&useron.expire);
+			return((int32_t *)&useron.expire);
 		case 0x89c91dc8:
-			return((long *)&useron.pwmod);
+			return((int32_t *)&useron.pwmod);
 		case 0x5b0d0c54:
-			return((long *)&useron.ns_time);
+			return((int32_t *)&useron.ns_time);
 
 		case 0xae256560:
-			return((long *)&cur_rate);
+			return((int32_t *)&cur_rate);
 		case 0x2b3c257f:
-			return((long *)&cur_cps);
+			return((int32_t *)&cur_cps);
 		case 0x1c4455ee:
-			return((long *)&dte_rate);
+			return((int32_t *)&dte_rate);
 		case 0x7fbf958e:
-			return((long *)&lncntr);
+			return((int32_t *)&lncntr);
 		case 0x5c1c1500:
-			return((long *)&tos);
+			return((int32_t *)&tos);
 		case 0x613b690e:
-			return((long *)&rows);
+			return((int32_t *)&rows);
 		case 0x205ace36:
-			return((long *)&autoterm);
+			return((int32_t *)&autoterm);
 		case 0x7d0ed0d1:
-			return((long *)&console);
+			return((int32_t *)&console);
 		case 0xbf31a280:
-			return((long *)&answertime);
+			return((int32_t *)&answertime);
 		case 0x83aa2a6a:
-			return((long *)&logontime);
+			return((int32_t *)&logontime);
 		case 0xb50cb889:
-			return((long *)&ns_time);
+			return((int32_t *)&ns_time);
 		case 0xae92d249:
-			return((long *)&last_ns_time);
+			return((int32_t *)&last_ns_time);
 		case 0x97f99eef:
-			return((long *)&online);
+			return((int32_t *)&online);
 		case 0x381d3c2a:
-			return((long *)&sys_status);
+			return((int32_t *)&sys_status);
 		case 0x7e29c819:
-			return((long *)&cfg.sys_misc);
+			return((int32_t *)&cfg.sys_misc);
 		case 0x11c83294:
-			return((long *)&cfg.sys_psnum);
+			return((int32_t *)&cfg.sys_psnum);
 		case 0x02408dc5:
 			sysvar_l[sysvar_li]=sys_timezone(&cfg);
 			break;
@@ -296,7 +296,7 @@ int32_t * sbbs_t::getintvar(csi_t *bin, int32_t name)
 			sysvar_l[sysvar_li]=cfg.node_scrnblank;
 			break;
 		case 0x709c07da:
-			return((long *)&cfg.node_misc);
+			return((int32_t *)&cfg.node_misc);
 		case 0xb17e7914:
 			sysvar_l[sysvar_li]=cfg.node_valuser;
 			break;
@@ -316,26 +316,26 @@ int32_t * sbbs_t::getintvar(csi_t *bin, int32_t name)
 			sysvar_l[sysvar_li]=cfg.new_level;
 			break;
 		case 0xccfe7c5d:
-			return((long *)&cfg.new_flags1);
+			return((int32_t *)&cfg.new_flags1);
 		case 0x55f72de7:
-			return((long *)&cfg.new_flags2);
+			return((int32_t *)&cfg.new_flags2);
 		case 0x22f01d71:
-			return((long *)&cfg.new_flags3);
+			return((int32_t *)&cfg.new_flags3);
 		case 0xbc9488d2:
-			return((long *)&cfg.new_flags4);
+			return((int32_t *)&cfg.new_flags4);
 		case 0x4b0aeb24:
-			return((long *)&cfg.new_exempt);
+			return((int32_t *)&cfg.new_exempt);
 		case 0x20cb6325:
-			return((long *)&cfg.new_rest);
+			return((int32_t *)&cfg.new_rest);
 		case 0x31178ba2:
-			return((long *)&cfg.new_cdt);
+			return((int32_t *)&cfg.new_cdt);
 		case 0x7345219f:
-			return((long *)&cfg.new_min);
+			return((int32_t *)&cfg.new_min);
 		case 0xb3f64be4:
 			sysvar_l[sysvar_li]=cfg.new_shell;
 			break;
 		case 0xa278584f:
-			return((long *)&cfg.new_misc);
+			return((int32_t *)&cfg.new_misc);
 		case 0x7342a625:
 			sysvar_l[sysvar_li]=cfg.new_expire;
 			break;
@@ -346,17 +346,17 @@ int32_t * sbbs_t::getintvar(csi_t *bin, int32_t name)
 			sysvar_l[sysvar_li]=cfg.expired_level;
 			break;
 		case 0x89b69753:
-			return((long *)&cfg.expired_flags1);
+			return((int32_t *)&cfg.expired_flags1);
 		case 0x10bfc6e9:
-			return((long *)&cfg.expired_flags2);
+			return((int32_t *)&cfg.expired_flags2);
 		case 0x67b8f67f:
-			return((long *)&cfg.expired_flags3);
+			return((int32_t *)&cfg.expired_flags3);
 		case 0xf9dc63dc:
-			return((long *)&cfg.expired_flags4);
+			return((int32_t *)&cfg.expired_flags4);
 		case 0x0e42002a:
-			return((long *)&cfg.expired_exempt);
+			return((int32_t *)&cfg.expired_exempt);
 		case 0x4569c62e:
-			return((long *)&cfg.expired_rest);
+			return((int32_t *)&cfg.expired_rest);
 		case 0xfcf3542e:
 			sysvar_l[sysvar_li]=cfg.min_dspace;
 			break;
@@ -364,7 +364,7 @@ int32_t * sbbs_t::getintvar(csi_t *bin, int32_t name)
 			sysvar_l[sysvar_li]=cfg.cdt_min_value;
 			break;
 		case 0xfcb5b274:
-			return((long *)&cfg.cdt_per_dollar);
+			return((int32_t *)&cfg.cdt_per_dollar);
 		case 0x4db200d2:
 			sysvar_l[sysvar_li]=cfg.leech_pct;
 			break;
@@ -372,31 +372,31 @@ int32_t * sbbs_t::getintvar(csi_t *bin, int32_t name)
 			sysvar_l[sysvar_li]=cfg.leech_sec;
 			break;
 		case 0x396b7167:
-			return((long *)&cfg.netmail_cost);
+			return((int32_t *)&cfg.netmail_cost);
 		case 0x5eeaff21:
 			sysvar_l[sysvar_li]=cfg.netmail_misc;
 			break;
 		case 0x82d9484e:
-			return((long *)&cfg.inetmail_cost);
+			return((int32_t *)&cfg.inetmail_cost);
 		case 0xe558c608:
-			return((long *)&cfg.inetmail_misc);
+			return((int32_t *)&cfg.inetmail_misc);
 
 		case 0xc6e8539d:
-			return((long *)&logon_ulb);
+			return((int32_t *)&logon_ulb);
 		case 0xdb0c9ada:
-			return((long *)&logon_dlb);
+			return((int32_t *)&logon_dlb);
 		case 0xac58736f:
-			return((long *)&logon_uls);
+			return((int32_t *)&logon_uls);
 		case 0xb1bcba28:
-			return((long *)&logon_dls);
+			return((int32_t *)&logon_dls);
 		case 0x9c5051c9:
-			return((long *)&logon_posts);
+			return((int32_t *)&logon_posts);
 		case 0xc82ba467:
-			return((long *)&logon_emails);
+			return((int32_t *)&logon_emails);
 		case 0x8e395209:
-			return((long *)&logon_fbacks);
+			return((int32_t *)&logon_fbacks);
 		case 0x8b12ba9d:
-			return((long *)&posts_read);
+			return((int32_t *)&posts_read);
 		case 0xe51c1956:
 			sysvar_l[sysvar_li]=(ulong)logfile_fp;
 			break;
@@ -420,12 +420,12 @@ int32_t * sbbs_t::getintvar(csi_t *bin, int32_t name)
 			break;
 
 		case 0x1e5052a7:
-			return((long *)&cfg.max_minutes);
+			return((int32_t *)&cfg.max_minutes);
 		case 0xedc643f1:
-			return((long *)&cfg.max_qwkmsgs);
+			return((int32_t *)&cfg.max_qwkmsgs);
 
 		case 0x430178ec:
-			return((long *)&cfg.uq);
+			return((int32_t *)&cfg.uq);
 
 		case 0x455CB929:
 			return(&bin->ftp_mode);
@@ -434,7 +434,7 @@ int32_t * sbbs_t::getintvar(csi_t *bin, int32_t name)
 			return(&bin->socket_error);
 
 		case 0xA0023A2E:
-			return((long *)&startup->options);
+			return((int32_t *)&startup->options);
 
 		case 0x16E2585F:
 			sysvar_l[sysvar_li]=client_socket;
